make the judgments tolerant for scrollTop, too.
I observed a case that page stucked on top, so add it.

in addition, enlarge the max scroll times from 10 to 20 to ensure it
will finish the scroll in faster computer, too.

ps.  in my laptop, it needs 6-7 iteration to finish a scroll. an
iteration took about 16-18ms
Win 10 64bit, i5-6300HQ, 8GB RAM, Firefox Nightly 57 (2017-08-13)